Rename institutional partnership pages
Old institutional-partners page goes away
Application page is under institutional-partnership-application
New IPP page is institutional-partnership
