--- src/app/pages/institutional-partner-program/institutional-partner-program.scss
@@ -1,81 +1,0 @@
-@import 'pattern-library/core/pattern-library/headers';
-
-$near-white: #f8f8f8;
-
-.institutional-partner-program.page {
-
-    section {
-        max-width: 100%;
-        overflow: auto;
-    }
-
-    .content-block {
-        @extend %content;
-
-        display: grid;
-        grid-gap: 4rem;
-        // max-width: 100%;
-
-        > * {
-            margin: 0;
-        }
-    }
-
-    .btn.primary {
-        @extend %button;
-        @extend %primary;
-
-        justify-self: left;
-    }
-
-    .near-white {
-        background-color: $near-white;
-    }
-
-    .white {
-        background-color: ui-color(white);
-    }
-
-    .green {
-        background-color: os-color(deep-green);
-        color: text-color(white);
-    }
-
-    .sticky-footer {
-        background-color: $near-white;
-        max-height: none;
-        padding: 1.5rem 3rem;
-
-        .button-group {
-            @extend %content;
-            display: grid;
-            justify-content: center;
-
-            @include wider-than($phone-max) {
-                grid-template: 'desc btn';
-                justify-content: space-between;
-
-                .btn {
-                    grid-area: btn;
-                }
-
-                .description {
-                    grid-area: desc;
-                }
-            }
-
-            .btn.medium {
-                @extend %button;
-                @extend %primary;
-
-                height: 5rem;
-                min-width: 17rem;
-            }
-
-            .description {
-                @include set-font(body-large);
-            }
-        }
-    }
-
-}
