+++ src/app/pages/institutional-partnership/sections/about/about.scss
@@ -1,6 +1,6 @@
 @import 'pattern-library/core/pattern-library/headers';
 
-.institutional-partner-program .about {
+.institutional-partnership .about {
     padding: 24rem 0 4rem;
     position: relative;
