+++ src/app/pages/institutional-partnership/sections/banner/banner.js
