+++ src/app/pages/institutional-partnership/sections/banner/banner.scss
@@ -1,6 +1,6 @@
 @import 'pattern-library/core/pattern-library/headers';
 
-.institutional-partner-program .banner {
+.institutional-partnership .banner {
     color: text-color(white);
     padding: 16rem 0 24rem;
     position: relative;
