+++ src/app/pages/institutional-partnership/sections/big-quote/big-quote.scss
