+++ src/app/pages/institutional-partnership/sections/overlapping-quote/overlapping-quote.html
