+++ src/app/pages/institutional-partnership/sections/overlapping-quote/overlapping-quote.scss
@@ -1,6 +1,6 @@
 @import 'pattern-library/core/pattern-library/headers';
 
-.institutional-partner-program.page .overlapping-quote {
+.institutional-partnership.page .overlapping-quote {
 
     overflow: visible;
     position: relative;
