+++ src/app/pages/institutional-partnership/sections/participants/established-partners/established-partners.scss
