+++ src/app/pages/institutional-partnership/sections/participants/participants.scss
@@ -1,6 +1,6 @@
 @import 'pattern-library/core/pattern-library/headers';
 
-.institutional-partner-program .participants {
+.institutional-partnership .participants {
     padding: 16rem 0 24rem;
 
     @include width-up-to($tablet-max) {
