+++ src/app/pages/institutional-partnership/sections/promoting/promoting.scss
@@ -1,6 +1,6 @@
 @import 'pattern-library/core/pattern-library/headers';
 
-.institutional-partner-program .promoting {
+.institutional-partnership .promoting {
     padding: 16rem 0;
     position: relative;
