+++ src/app/pages/institutional-partnership/sections/results/results.scss
@@ -1,6 +1,6 @@
 @import 'pattern-library/core/pattern-library/headers';
 
-.institutional-partner-program .results {
+.institutional-partnership .results {
     padding: 16rem 0 24rem;
     position: relative;
