+++ src/app/pages/institutional-partnership/sections/sign-up/sign-up.scss
