+++ src/app/pages/institutional-partnership/sections/small-quote/small-quote.scss
