+++ src/app/pages/institutional-partnership/sections/speaking/speaking.scss
@@ -1,6 +1,6 @@
 @import 'pattern-library/core/pattern-library/headers';
 
-.institutional-partner-program .speaking {
+.institutional-partnership .speaking {
     padding: 12rem 0 10rem;
     position: relative;
